+++ libs/atqr/domain/src/lib/EmailService/index.ts
@@ -1,4 +1,3 @@
 export * from './email.service';
 export * from './email.interface';
-export * from './emailAddress';
 export * from '../EmailService/templates/';
