+++ libs/atqr/domain/src/lib/domain-spec/TODO.md
@@ -1,0 +1,7 @@
+# ATQR Domain Spec Tasks
+
+- [x] [AcceptChallenge?](api.disl) - @GabrielaRsanches
+- [ ] Adicionar Gender ao User (M, F, Não-binário)
+- [ ] O jogador que se inscreve na challenge por link escolhe seu Juiz ou não?
+- [ ] Em challenges com inscrição por link, como definir o número máximo de jogadores por Juiz?
+- [ ] Criar serviço EmailBlacklist que armazena emails de pessoas que pediram pra serem removidos da lista
