+++ libs/atqr/domain/src/lib/domain-spec/enums.disl
@@ -1,4 +1,4 @@
-context TheGame
+context Challenges
   enum ChallengeStatus
     Unverified
     /**
@@ -18,35 +18,45 @@ context TheGame
     Finished
 
   enum PlayerOutcome
+    StillPlaying
     Succeeded
     Failed
-
-  enum PaymentStatus
-    NotNeeded
-    Abandoned
-    Failed
-    Paid
-    Pending
-    Refunded
-    RefundRequested
-    Defaulted
-
-  enum PaymentMethodType
-    CreditCard
-    DebitCard
 
   enum PaymentFundsStatus
     Unverified
     Authorized
     Denied
 
-  enum SupervisionStatus
+  /**
+   * The status of a challenge payment.
+   * Here's
+   */
+  enum PaymentStatus
+    NotDueYet
+    Defaulted
+    Failed
+    Forgiven
+    Paid
+    Pending
+    Refunded
+    RefundRequested
+
+  enum PaymentMethodType
+    CreditCard
+    DebitCard
+
+  enum ParticipationStatus
     NotRequested
     Requested
     Accepted
     Rejected
 
-  enum ChallengeJudgement
+  enum JudgementStatus
+    NotDueYet
     NotRequested
     Requested
     Judged
+
+  enum ParticipationRole
+    Player
+    Judge
