--- libs/atqr/domain/src/lib/domain-spec/services.disl
@@ -1,198 +1,0 @@
-context TheGame
-  service  StripeService
-    requires configService: ConfigService
-
-    async retrieveCustomer(
-      customerId: string
-    ): Promise<Stripe.Response<Stripe.Customer | Stripe.DeletedCustomer>> {
-      const customer = await this.stripeClient.customers.retrieve(customerId);
-      if (customer.deleted) {
-        throw new Error('User was deleted');
-      }
-      return customer;
-    }
-
-    // TODO: Provide parameter type
-    async getPaymentMethod(customerId): Promise<Stripe.PaymentMethod> {
-      const paymentMethods = await this.stripeClient.paymentMethods.list({
-        customer: customerId,
-        type: 'card',
-      });
-      return paymentMethods.data[0];
-    }
-
-    async createSetupIntent(
-      challengedId?: string
-    ): Promise<Stripe.Response<Stripe.SetupIntent>> {
-      const customer = await this.stripeClient.customers.create(
-        challengedId && {
-          metadata: {
-            challengedId,
-          },
-        }
-      );
-
-      return await this.stripeClient.setupIntents.create({
-        customer: customer.id,
-        payment_method_types: ['card'],
-      });
-    }
-
-    // TODO: Provide parameter type
-    async chargeCard(customerId) {
-      // TODO: Define Payment intent and Customer Types
-      let paymentIntent, customer;
-      try {
-        // You need to attach the PaymentMethod to a Customer in order to reuse
-        // Since we are using test cards, create a new Customer here
-        // You would do this in your payment flow that saves cards
-        customer = await this.retrieveCustomer(customerId);
-
-        // List the customer's payment methods to find one to charge
-        const paymentMethod = await this.getPaymentMethod(customerId);
-
-        // Create and confirm a PaymentIntent with the order amount, currency,
-        // Customer and PaymentMethod ID
-
-        paymentIntent = this.createPaymentIntent(
-          customer,
-          paymentMethod,
-          1499, // Todo: Get value from challenge,
-          false
-        );
-
-        return {
-          succeeded: true,
-          clientSecret: paymentIntent.client_secret,
-          publicKey: this.publicKey,
-        };
-      } catch (err) {
-        if (err.code === 'authentication_required') {
-          // Bring the customer back on-session to authenticate the purchase
-          // You can do this by sending an email or app notification to let them know
-          // the off-session purchase failed
-          // Use the PM ID and client_secret to authenticate the purchase
-          // without asking your customers to re-enter their details
-          return {
-            error: 'authentication_required',
-            paymentMethod: err.raw.payment_method.id,
-            clientSecret: err.raw.payment_intent.client_secret,
-            publicKey: process.env.STRIPE_PUBLISHABLE_KEY,
-            amount: 1499, // TODO: Get value from challenge
-            card: {
-              brand: err.raw.payment_method.card.brand,
-              last4: err.raw.payment_method.card.last4,
-            },
-          };
-        } else if (err.code) {
-          // The card was declined for other reasons (e.g. insufficient funds)
-          // Bring the customer back on-session to ask them for a new payment method
-          return {
-            error: err.code,
-            clientSecret: err.raw.payment_intent.client_secret,
-            publicKey: process.env.STRIPE_PUBLISHABLE_KEY,
-          };
-        } else {
-          console.log('Unknown error occurred', err);
-        }
-      }
-      return;
-    }
-
-    // TODO: Provide parameter type
-    async verifyFunds(customerId) {
-      // TODO: Define Payment intent and Customer Types
-      let paymentIntent, customer;
-      try {
-        // You need to attach the PaymentMethod to a Customer in order to reuse
-        // Since we are using test cards, create a new Customer here
-        // You would do this in your payment flow that saves cards
-        customer = await this.retrieveCustomer(customerId);
-
-        // List the customer's payment methods to find one to charge
-        const paymentMethod = await this.getPaymentMethod(customerId);
-
-        // Create and confirm a PaymentIntent with the order amount, currency,
-        // Customer and PaymentMethod ID
-
-        paymentIntent = this.createPaymentIntent(
-          customer,
-          paymentMethod,
-          1499, // Todo: Get value from challenge
-          false
-        );
-
-        return {
-          succeeded: true,
-          clientSecret: paymentIntent.client_secret,
-          publicKey: this.publicKey,
-        };
-      } catch (err) {
-        if (err.code === 'authentication_required') {
-          // Bring the customer back on-session to authenticate the purchase
-          // You can do this by sending an email or app notification to let them know
-          // the off-session purchase failed
-          // Use the PM ID and client_secret to authenticate the purchase
-          // without asking your customers to re-enter their details
-          return {
-            error: 'authentication_required',
-            paymentMethod: err.raw.payment_method.id,
-            clientSecret: err.raw.payment_intent.client_secret,
-            publicKey: process.env.STRIPE_PUBLISHABLE_KEY,
-            amount: 1499, // TODO: Get value from challenge
-            card: {
-              brand: err.raw.payment_method.card.brand,
-              last4: err.raw.payment_method.card.last4,
-            },
-          };
-        } else if (err.code) {
-          // The card was declined for other reasons (e.g. insufficient funds)
-          // Bring the customer back on-session to ask them for a new payment method
-          return {
-            error: err.code,
-            clientSecret: err.raw.payment_intent.client_secret,
-            publicKey: process.env.STRIPE_PUBLISHABLE_KEY,
-          };
-        } else {
-          console.log('Unknown error occurred', err);
-        }
-      }
-    }
-
-    private async createPaymentIntent(
-      customer: Stripe.Customer,
-      paymentMethod: Stripe.PaymentMethod,
-      amount: number,
-      confirm = false
-    ): Promise<Stripe.Response<Stripe.PaymentIntent>> {
-      return await this.stripeClient.paymentIntents.create({
-        amount,
-        currency: 'brl',
-        payment_method: paymentMethod.id,
-        customer: customer.id,
-        off_session: true,
-        confirm,
-      });
-    }
-
-    constructEvent(evtBody, signature): Stripe.Event {
-      console.log('Constructing Event');
-      return this.stripeClient.webhooks.constructEvent(
-        evtBody,
-        signature,
-        this.endpointSecret
-      );
-    }
-
-    // Todo: Implement webhook method:
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    handlePaymentIntentSucceeded(paymentIntent: Stripe.PaymentIntent) {
-      throw new Error('Method not implemented.');
-    }
-
-    // Todo: Implement webhook method:
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    handlePaymentMethodAttached(paymentMethod: Stripe.PaymentMethod) {
-      throw new Error('Method not implemented.');
-    }
-  }
