+++ package.json
@@ -59,7 +59,6 @@
     "node-ts": "5.1.2",
     "nodemailer": "6.7.3",
     "nodemailer-mailgun-transport": "2.1.3",
-    "parcel": "2.8.0",
     "prisma": "4.1.0",
     "pug": "3.0.2",
     "pug-cli": "1.0.0-alpha6",
@@ -112,7 +111,7 @@
     "jest": "27.5.1",
     "jest-preset-angular": "11.0.0",
     "os-browserify": "^0.3.0",
-    "parcel": "2.7.0",
+    "parcel": "2.8.0",
     "path-browserify": "^1.0.1",
     "postcss": "8.4.5",
     "prettier": "2.3.1",
